LayoutTests/imported/w3c/web-platform-tests/intersection-observer/v2/blur-filter-expected.txt


PASS IntersectionObserverV2 in a single document using the implicit root, with an occluding element.
PASS First rAF.
FAIL occluder.style.opacity = 0 assert_equals: entries.length expected 2 but got 1
